+++ README.md
@@ -315,10 +315,9 @@ After the code has been generated by ScalaPB, you can use wookiee-grpc for servi
 ```scala
 import java.lang.Thread.UncaughtExceptionHandler
 import java.util.concurrent.{Executors, ForkJoinPool, ThreadFactory}
-
 import cats.effect.IO
+import com.oracle.infy.wookiee.grpc.model.Host
 import com.oracle.infy.wookiee.grpc.{WookieeGrpcChannel, WookieeGrpcServer}
-import com.oracle.infy.wookiee.model.Host
 // This is from ScalaPB generated code
 import com.oracle.infy.wookiee.myService.MyServiceGrpc.MyService
 import com.oracle.infy.wookiee.myService.{HelloRequest, HelloResponse, MyServiceGrpc}
