+++ VERSION
@@ -1,1 +1,1 @@
-2.2.6
+2.2.7
