+++ src/test/scala/com/oracle/infy/wookiee/grpc/IntegrationConstable.scala
@@ -7,9 +7,9 @@ import com.oracle.infy.wookiee.grpc.common.ConstableCommon
 import com.oracle.infy.wookiee.grpc.contract.ListenerContract
 import com.oracle.infy.wookiee.grpc.impl.{Fs2CloseableImpl, WookieeGrpcHostListener, ZookeeperHostnameService}
 import com.oracle.infy.wookiee.grpc.json.HostSerde
+import com.oracle.infy.wookiee.grpc.model.Host
 import com.oracle.infy.wookiee.grpc.tests.{GrpcListenerTest, GrpcLoadBalanceTest}
-import com.oracle.infy.wookiee.model.Host
-import com.oracle.infy.wookiee.utils.implicits._
+import com.oracle.infy.wookiee.grpc.utils.implicits._
 import fs2.Stream
 import fs2.concurrent.Queue
 import org.typelevel.log4cats.Logger
