+++ src/test/scala/com/oracle/infy/wookiee/grpc/UnitTestConstable.scala
@@ -6,8 +6,8 @@ import cats.implicits._
 import com.oracle.infy.wookiee.grpc.common.ConstableCommon
 import com.oracle.infy.wookiee.grpc.contract.ListenerContract
 import com.oracle.infy.wookiee.grpc.impl.{Fs2CloseableImpl, MockHostNameService, WookieeGrpcHostListener}
+import com.oracle.infy.wookiee.grpc.model.Host
 import com.oracle.infy.wookiee.grpc.tests.{GrpcListenerTest, SerdeTest}
-import com.oracle.infy.wookiee.model.Host
 import fs2.Stream
 import fs2.concurrent.Queue
 import org.typelevel.log4cats.noop.NoOpLogger
