+++ src/test/scala/com/oracle/infy/wookiee/grpc/tests/GrpcMultipleClientsTest.scala
@@ -2,15 +2,15 @@ package com.oracle.infy.wookiee.grpc.tests
 
 import cats.effect.{Blocker, ContextShift, IO, Timer}
 import com.oracle.infy.wookiee.grpc.common.UTestScalaCheck
+import com.oracle.infy.wookiee.grpc.model.{Host, HostMetadata}
 import com.oracle.infy.wookiee.grpc.settings.{ChannelSettings, ServerSettings}
 import com.oracle.infy.wookiee.grpc.{WookieeGrpcChannel, WookieeGrpcServer, ZookeeperUtils}
-import com.oracle.infy.wookiee.model.LoadBalancers.RoundRobinPolicy
-import com.oracle.infy.wookiee.model.{Host, HostMetadata}
+import com.oracle.infy.wookiee.grpc.model.LoadBalancers.RoundRobinPolicy
 import com.oracle.infy.wookiee.myService.MyServiceGrpc.MyService
 import com.oracle.infy.wookiee.myService.{HelloRequest, HelloResponse, MyServiceGrpc}
 import com.oracle.infy.wookiee.myService2.MyService2Grpc.MyService2
 import com.oracle.infy.wookiee.myService2.{HelloRequest2, HelloResponse2, MyService2Grpc}
-import com.oracle.infy.wookiee.utils.implicits.MultiversalEquality
+import com.oracle.infy.wookiee.grpc.utils.implicits.MultiversalEquality
 import org.typelevel.log4cats.Logger
 import io.grpc.ServerServiceDefinition
 import org.apache.curator.framework.CuratorFramework
