+++ src/test/scala/com/oracle/infy/wookiee/grpc/tests/SerdeTest.scala
@@ -3,8 +3,8 @@ package com.oracle.infy.wookiee.grpc.tests
 import cats.implicits.{catsSyntaxEq => _}
 import com.oracle.infy.wookiee.grpc.common.{HostGenerator, UTestScalaCheck}
 import com.oracle.infy.wookiee.grpc.json.HostSerde._
-import com.oracle.infy.wookiee.model.Host
-import com.oracle.infy.wookiee.utils.implicits._
+import com.oracle.infy.wookiee.grpc.model.Host
+import com.oracle.infy.wookiee.grpc.utils.implicits._
 import org.scalacheck.Prop
 import org.scalacheck.Prop.forAll
 import utest.{Tests, test}
