+++ src/test/scala/com/oracle/infy/wookiee/grpcdev/tests/GrpcDevTest.scala
@@ -2,7 +2,7 @@ package com.oracle.infy.wookiee.grpcdev.tests
 
 import com.oracle.infy.wookiee.grpc.srcgen.SourceGenModel.ScalaTextSource
 import com.oracle.infy.wookiee.grpcdev.common.TestModel._
-import com.oracle.infy.wookiee.utils.implicits._
+import com.oracle.infy.wookiee.grpc.utils.implicits._
 import utest.{Tests, test}
 
 object GrpcDevTest {
