+++ src/test/scala/com/oracle/infy/wookiee/grpcdev/tests/SrcGenTest.scala
@@ -2,7 +2,7 @@ package com.oracle.infy.wookiee.grpcdev.tests
 
 import com.oracle.infy.wookiee.srcgen.implicits._
 import com.oracle.infy.wookiee.srcgen.Example._
-import com.oracle.infy.wookiee.utils.implicits._
+import com.oracle.infy.wookiee.grpc.utils.implicits._
 import cats.implicits._
 import utest.{Tests, test}
