+++ src/test/scala/com/oracle/infy/wookiee/health/tests/HealthRoutesTest.scala
@@ -5,7 +5,7 @@ import com.oracle.infy.wookiee.grpc.common.UTestScalaCheck
 import com.oracle.infy.wookiee.health.HeathCheckServer
 import com.oracle.infy.wookiee.health.json.Serde
 import com.oracle.infy.wookiee.health.model.{Critical, Normal, State, WookieeHealth}
-import com.oracle.infy.wookiee.utils.implicits._
+import com.oracle.infy.wookiee.grpc.utils.implicits._
 import org.http4s._
 import org.http4s.implicits._
 import utest.{Tests, test}
