+++ src/test/scala/com/oracle/infy/wookiee/metrics/tests/MetricsServiceTest.scala
@@ -5,7 +5,7 @@ import com.oracle.infy.wookiee.functional.metrics.WookieeMetricsService
 import com.oracle.infy.wookiee.functional.metrics.core.WookieeMetricsReporter
 import com.oracle.infy.wookiee.grpc.common.UTestScalaCheck
 import com.oracle.infy.wookiee.metrics.model.WookieeRegistry
-import com.oracle.infy.wookiee.utils.implicits._
+import com.oracle.infy.wookiee.grpc.utils.implicits._
 import utest.{Tests, test}
 
 import java.util.concurrent.TimeUnit
