+++ wookiee-grpc/src/main/scala/com/oracle/infy/wookiee/grpc/contract/HostnameServiceContract.scala
@@ -2,7 +2,7 @@ package com.oracle.infy.wookiee.grpc.contract
 
 import cats.data.EitherT
 import com.oracle.infy.wookiee.grpc.errors.Errors.WookieeGrpcError
-import com.oracle.infy.wookiee.model.Host
+import com.oracle.infy.wookiee.grpc.model.Host
 
 trait HostnameServiceContract[F[_], S[_[_], _]] {
   def shutdown: EitherT[F, WookieeGrpcError, Unit]
