+++ wookiee-grpc/src/main/scala/com/oracle/infy/wookiee/grpc/model/Host.scala
@@ -1,4 +1,4 @@
-package com.oracle.infy.wookiee.model
+package com.oracle.infy.wookiee.grpc.model
 
 case class HostMetadata(load: Int, quarantined: Boolean)
 final case class Host(version: Long, address: String, port: Int, metadata: HostMetadata)
