+++ wookiee-grpc/src/main/scala/com/oracle/infy/wookiee/grpc/model/LoadBalancers.scala
@@ -1,12 +1,14 @@
-package com.oracle.infy.wookiee.model
+package com.oracle.infy.wookiee.grpc.model
 
 object LoadBalancers {
 
   sealed trait LoadBalancingPolicy
 
   sealed trait RoundRobinPolicy extends LoadBalancingPolicy
+
   final case object RoundRobinPolicy extends RoundRobinPolicy
 
   sealed trait RoundRobinWeightedPolicy extends LoadBalancingPolicy
+
   final case object RoundRobinWeightedPolicy extends RoundRobinWeightedPolicy
 }
